Arraylists - static api for animations
