+++ src/structvisualizer/animatecollections/animatemethods/arraylist/Add.java
@@ -4,16 +4,14 @@ import javafx.animation.FadeTransition;
 import javafx.animation.TranslateTransition;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.StackPane;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Rectangle;
-import javafx.scene.text.Text;
 import javafx.util.Duration;
 import parser.SomeClass;
 import structvisualizer.animatecollections.animatemethods.AnimateMethod;
 import structvisualizer.data.OutputStrings;
-import structvisualizer.valuefactories.TypeValueFactory;
 
 import java.util.ArrayList;
+
+import static structvisualizer.animatecollections.animatemethods.arraylist.Arraylists.addToArrayList;
 
 /**
  * Animates ArrayList.add() on Pane
@@ -27,6 +25,7 @@ public class Add extends AnimateMethod {
      * @param canvasPane
      * @param type
      * @param obj
+     *
      * @see AnimateMethod
      */
     public Add(Pane canvasPane, String type, SomeClass obj) {
@@ -34,41 +33,17 @@ public class Add extends AnimateMethod {
 
     }
 
-    /**
-     * Static method to create ArrayList<StackPane> of elements to show on Pane (canvas)
-     *
-     * @param rectangles ArrayList where to add StackPane
-     * @param type       Type of values in ArrayList
-     * @return ArrayList<StackPane>
-     */
-    public static ArrayList<StackPane> addToArrayList(ArrayList<StackPane> rectangles, String type) {
-        StackPane stack = new StackPane();
-        stack.setLayoutX(data.getStackPaneSize());
-        stack.setLayoutY(data.getStackPaneSize());
-
-        rectangles.add(data.getNumOfStackPanes(), stack);
-        Rectangle r = new Rectangle();
-        r.setWidth(data.getWidth());
-        r.setHeight(data.getHeight());
-        r.setFill(Color.TRANSPARENT);
-        r.setStroke(Color.BLACK);
-
-        Text text = new Text(TypeValueFactory.get(type, data.getNumOfStackPanes()));
-
-        rectangles.get(data.getNumOfStackPanes()).getChildren().addAll(r, text);
-        return rectangles;
-    }
-
 
     /**
      * @param type string with type of values in ArrayList
+     *
      * @see AnimateMethod
      */
     @Override
     public void animate(String type) {
 
 
-        ArrayList<StackPane> rectangles = Construct.createArrayList(type);
+        ArrayList<StackPane> rectangles = Arraylists.createArrayList(type);
 
         rectangles = addToArrayList(rectangles, type);
 
@@ -89,6 +64,7 @@ public class Add extends AnimateMethod {
 
     /**
      * @param os
+     *
      * @return
      * @see AnimateMethod
      */
