+++ src/structvisualizer/animatecollections/animatemethods/arraylist/Clear.java
@@ -7,15 +7,15 @@ package structvisualizer.animatecollections.animatemethods.arraylist;
  *
  */
 
-import javafx.animation.FadeTransition;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.StackPane;
-import javafx.util.Duration;
 import parser.SomeClass;
 import structvisualizer.animatecollections.animatemethods.AnimateMethod;
 import structvisualizer.data.OutputStrings;
 
 import java.util.ArrayList;
+
+import static structvisualizer.animatecollections.animatemethods.arraylist.Arraylists.clearAnimation;
 
 /**
  * <what class do>
@@ -37,28 +37,17 @@ public class Clear extends AnimateMethod {
         super(canvasPane, type, obj);
     }
 
-    public static ArrayList<StackPane> clearAnimation(ArrayList<StackPane> stack) {
-        int counter = 1000;
-        for(StackPane s : stack) {
-            FadeTransition fade = new FadeTransition(Duration.millis(data.getFade() + counter), s);
-            fade.setFromValue(1);
-            fade.setToValue(0);
-            fade.play();
-            counter -= 100;
-
-        }
-        return stack;
-    }
 
     @Override
     public void animate(String type) throws UnsupportedOperationException {
-        ArrayList<StackPane> stack = Construct.createArrayList(type);
+        ArrayList<StackPane> stack = Arraylists.createArrayList(type);
         canvasPane.getChildren().addAll(stack);
         clearAnimation(stack);
     }
 
     /**
      * @param os Class with templates of result
+     *
      * @return
      * @see AnimateMethod
      */
