+++ src/structvisualizer/animatecollections/animatemethods/arraylist/Get.java
@@ -13,8 +13,6 @@ import parser.SomeClass;
 import structvisualizer.animatecollections.animatemethods.AnimateMethod;
 import structvisualizer.data.OutputStrings;
 
-import static structvisualizer.animatecollections.animatemethods.arraylist.Construct.createArrayList;
-
 /**
  * <what class do>
  *
@@ -41,23 +39,23 @@ public class Get extends AnimateMethod implements MethodsForSearch {
     @Override
     public void animate(String type) throws UnsupportedOperationException {
 
-//        ArrayList<StackPane> rectangles = createArrayList(type);
-//        canvasPane.getChildren().addAll(rectangles);
-//        AnimateMethod.setTooltip(rectangles, type, customClass);
-//
-//        StackPane sp = newStackPane(0, 0, canvasPane, data.getHightlightColor());
-//
-//        double finalX = 0;
-//        if(type == Types.INT) finalX += data.getWidth();
-//        for(int i = 0; i < data.getNumOfStackPanes(); i++) {
-//            String text = rectangles.get(i).getChildren().get(1).toString();
-//            // indexOf - VALUE OF ELEMENT WHAT WE LOOKING FOR
-//            if(!text.substring(11, 12).equals(indexOf)) {
-//                finalX += data.getWidth();
-//            }
-//        }
-//        animateSearch(sp, 0, 0, finalX, 0, canvasPane);
-        Set.searchForElement(this, type, canvasPane, customClass, indexOf, index);
+        //        ArrayList<StackPane> rectangles = createArrayList(type);
+        //        canvasPane.getChildren().addAll(rectangles);
+        //        AnimateMethod.setTooltip(rectangles, type, customClass);
+        //
+        //        StackPane sp = newStackPane(0, 0, canvasPane, data.getHightlightColor());
+        //
+        //        double finalX = 0;
+        //        if(type == Types.INT) finalX += data.getWidth();
+        //        for(int i = 0; i < data.getNumOfStackPanes(); i++) {
+        //            String text = rectangles.get(i).getChildren().get(1).toString();
+        //            // indexOf - VALUE OF ELEMENT WHAT WE LOOKING FOR
+        //            if(!text.substring(11, 12).equals(indexOf)) {
+        //                finalX += data.getWidth();
+        //            }
+        //        }
+        //        animateSearch(sp, 0, 0, finalX, 0, canvasPane);
+        Arraylists.searchForElement(this, type, canvasPane, customClass, indexOf, index);
     }
 
     @Override
